Move shared logic to abstract planet class.
