+++ lib/src/main/kotlin/Planet.kt
@@ -8,9 +8,7 @@ import kotlin.math.ceil
 import kotlin.math.max
 import kotlin.math.min
 
-class Planet constructor(val seed: Double = Defaults.SEED): AbstractPlanet() {
-    private val squishedSeed = squishSeed(seed)
-    private val tetra = Tetrahedron.buildDefault(squishedSeed)
+class Planet constructor(seed: Double = Defaults.SEED): AbstractPlanet(seed) {
     private var elevations: MapTileElevations
 
     init {
@@ -51,87 +49,11 @@ class Planet constructor(val seed: Double = Defaults.SEED): AbstractPlanet() {
         fun get(seed: Double): Planet {
             return planetCache.get(seed) { it -> Planet(it) }.get()!!
         }
-
-        private fun MutableList<Point>.partitionInPlaceBy(compareFunc: (Point) -> Boolean): Int {
-            // Sorts all records that return TRUE by the comparator to the front of the list, and
-            //   returns the index of the first record that returns FALSE.
-            var pointerOne = 0
-            var pointerTwo = this.size - 1
-            while (true) {
-                while (pointerOne < this.size && compareFunc(this[pointerOne])) {
-                    pointerOne++
-                }
-                while (pointerTwo >= 0 && !compareFunc(this[pointerTwo])) {
-                    pointerTwo--
-                }
-                if (pointerOne >= pointerTwo) {
-                    break
-                }
-                // Using .also{} to swap! https://stackoverflow.com/questions/45377802/swap-function-in-kotlin
-                this[pointerOne] = this[pointerTwo].also { this[pointerTwo] = this[pointerOne] }
-            }
-            return pointerOne
-        }
     }
     override fun getMapTile(mapTileKey: MapTileKey): MapTile {
         return mapTileCache.get(mapTileKey) { key -> buildMapTile(key, elevations) }.get()!!
     }
-
-    fun getElevationAt(lat: Double, lon: Double, resolution: Int): Point {
-        val point = Point.fromSpherical(lat = lat, lon = lon, resolution = resolution)
-        var current = this.tetra
-        var subdivisions = 0
-        while (current.longestSide > resolution){
-            subdivisions += 1
-            val (subOne, subTwo) = current.subdivide()
-            current = if (subOne.contains(point)){
-                subOne
-            } else {
-                assert(subTwo.contains(point))
-                subTwo
-            }
-        }
-        return point.copy(alt=current.averageAltitude)
-    }
-
-    fun getMultipleElevations(points: MutableList<Point>, current: Tetrahedron = this.tetra): MutableList<Point> {
-        if (points.isEmpty()){
-            return points
-        }
-
-        // Since some points may not require any further recursion, we can filter them out and set elevation.
-        val doneIndex: Int = points.partitionInPlaceBy { point ->  current.longestSide <= point.resolution }
-        for (i in 0 until doneIndex) {
-            points[i] = points[i].copy(alt=current.averageAltitude)
-        }
-
-        // If we've identified all points as done, no need to subdivide and recurse.
-        if (doneIndex == points.size) {
-            return points
-        }
-
-        // All remaining points must be sorted into the tetrahedron they are contained within.
-        val (leftTetra, rightTetra) = current.subdivide()
-
-        val pendingPoints = points.subList(doneIndex, points.size)
-        val containmentIndex = pendingPoints.partitionInPlaceBy { point ->
-            leftTetra.contains(point)
-        }
-
-        val leftPoints = pendingPoints.subList(0, containmentIndex)
-        val rightPoints = pendingPoints.subList(containmentIndex, pendingPoints.size)
-
-        getMultipleElevations(leftPoints, leftTetra)
-        getMultipleElevations(rightPoints, rightTetra)
-        return points
-    }
-
-    private fun buildMapTile(mapTileKey: MapTileKey, elevations: MapTileElevations): MapTile {
-        val pointsWithElevations = calculateMapTilePoints(mapTileKey)
-        return MapTile(mapTileKey, pointsWithElevations, elevations)
-    }
-
-    private fun calculateMapTilePoints(mapTileKey: MapTileKey): MutableList<Point> {
+    override fun calculateMapTilePoints(mapTileKey: MapTileKey): MutableList<Point> {
         val allPoints = ArrayList<Point>()
 
         val tileBounds: MapTileBounds = MapTileBounds.fromGeographicTileXYZ(mapTileKey.z, mapTileKey.x, mapTileKey.y)
